fix(huddle): stop 20 Hz speaker-level churn from re-rendering every useHuddle consumer
With a huddle open, the Rust playout loop emits huddle-speaker-levels every
50 ms (playout.rs SPEAKER_LEVEL_TICK_MS). Each event lands in fresh state, and
because HuddleProvider wraps the whole app and its context value was an inline
object literal, every useHuddle() consumer — including ChannelScreen and
message rows — re-rendered on every tick. Measured: a silent huddle drove a
sustained ~41 ChannelScreen renders/sec; suppressing only the speaker-level
setState dropped it to ~4/sec (A/B on the instrumented rig; receipts in the
buzz-conversational-agents thread).

Fix, minimal and DRY:
- split the high-frequency fields (micLevel, activeSpeakers, speakerLevels)
  into a separate HuddleLevelsContext consumed via useHuddleLevels() only by
  the meter components (HuddleBar, HuddleRoomHeader, HuddleProfileControl);
- memoize the main HuddleContext value so level churn (and any other provider
  re-render) no longer mints a new context identity for everyone else;
- extract the mic-level analyser into useMicLevelAnalyser so the level
  pipeline lives in one place (also keeps HuddleContext.tsx under the
  file-size ratchet).

Level meters keep their 20-30 Hz updates; the rest of the app only re-renders
when a value it actually consumes changes.
