+++ desktop/src/features/huddle/HuddleContext.types.ts
@@ -1,6 +1,17 @@
 import type { AudioInputDevice } from "./lib/useAudioDevices";
 import type { VoiceInputMode } from "./lib/useHuddlePttState";
 
+/**
+ * High-frequency audio-level fields, split from {@link HuddleContextValue} so
+ * their 20-30 Hz updates only re-render the meter components that consume
+ * them — not every `useHuddle()` consumer across the app.
+ */
+export interface HuddleLevelsValue {
+  micLevel: number;
+  activeSpeakers: string[];
+  speakerLevels: Record<string, number>;
+}
+
 export interface HuddleContextValue {
   localAudioTrack: MediaStreamTrack | null;
   isStarting: boolean;
@@ -11,12 +22,9 @@ export interface HuddleContextValue {
   toggleMute: () => void;
   /** Interrupt this agent only if it still owns the active utterance. */
   interruptAgentSpeech: (agentPubkey: string) => Promise<void>;
-  micLevel: number;
   pttActive: boolean;
   voiceInputMode: VoiceInputMode;
   setVoiceInputMode: (mode: VoiceInputMode) => Promise<void>;
-  activeSpeakers: string[];
-  speakerLevels: Record<string, number>;
   audioDevices: AudioInputDevice[];
   selectedDeviceId: string;
   setSelectedDeviceId: (id: string) => void;
